requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pymongo==4.10.1
neo4j==5.27.0
python-dotenv==1.0.1
pydantic==2.10.3
dnspython==2.7.0